requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
webdataset>=0.2.86
soundfile>=0.12.1
requests>=2.31.0
librosa>=0.10.0
jiwer>=3.0.3
pyyaml>=6.0.1
tqdm>=4.66.1
accelerate>=0.23.0
deepspeed>=0.16.1
tensorboard
faster-whisper>=1.1.1
sentencepiece>=0.2.0
TTS==0.22.0
